+++ src/Bridge/Doctrine/Orm/Filter/SearchFilter.php
@@ -21,8 +21,10 @@ use ApiPlatform\Core\Bridge\Doctrine\Orm\Util\QueryBuilderHelper;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Util\QueryNameGeneratorInterface;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
 use Doctrine\DBAL\Types\Type as DBALType;
+use Doctrine\ORM\Query\Parameter;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\Mapping\ClassMetadata;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\PropertyAccess\PropertyAccess;
@@ -113,7 +115,7 @@ class SearchFilter extends AbstractContextAwareFilter implements SearchFilterInt
                 $caseSensitive = false;
             }
 
-            $this->addWhereByStrategy($strategy, $queryBuilder, $queryNameGenerator, $alias, $field, $values, $caseSensitive);
+            $this->addWhereByStrategy($strategy, $queryBuilder, $queryNameGenerator, $alias, $field, $values, $caseSensitive, $metadata);
 
             return;
         }
@@ -149,15 +151,28 @@ class SearchFilter extends AbstractContextAwareFilter implements SearchFilterInt
             $associationField = $associationFieldIdentifier;
         }
 
+        $type = $metadata->getTypeOfField($associationField);
+
         if (1 === \count($values)) {
             $queryBuilder
                 ->andWhere($queryBuilder->expr()->eq($associationAlias.'.'.$associationField, ':'.$valueParameter))
-                ->setParameter($valueParameter, $values[0]);
-        } else {
-            $queryBuilder
-                ->andWhere($queryBuilder->expr()->in($associationAlias.'.'.$associationField, ':'.$valueParameter))
-                ->setParameter($valueParameter, $values);
+                ->setParameter($valueParameter, $values[0], $type);
+
+            return;
         }
+
+        $parameters = $queryBuilder->getParameters();
+        $inQuery = [];
+
+        foreach ($values as $val) {
+            $inQuery[] = ':'.$valueParameter;
+            $parameters->add(new Parameter($valueParameter, $val, $type));
+            $valueParameter = $queryNameGenerator->generateParameterName($associationField);
+        }
+
+        $queryBuilder
+            ->andWhere($associationAlias.'.'.$associationField.' IN ('.implode(', ', $inQuery).')')
+            ->setParameters($parameters);
     }
 
     /**
@@ -165,8 +180,15 @@ class SearchFilter extends AbstractContextAwareFilter implements SearchFilterInt
      *
      * @throws InvalidArgumentException If strategy does not exist
      */
-    protected function addWhereByStrategy(string $strategy, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $alias, string $field, $values, bool $caseSensitive)
+    protected function addWhereByStrategy(string $strategy, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $alias, string $field, $values, bool $caseSensitive/*, ClassMetadata $metadata*/)
     {
+        if (\func_num_args() > 7 && ($metadata = func_get_arg(7)) instanceof ClassMetadata) {
+            $type = $metadata->getTypeOfField($field);
+        } else {
+            @trigger_error(sprintf('Method %s() will have a 8th argument `$metadata` in version API Platform 3.0.', __FUNCTION__), \E_USER_DEPRECATED);
+            $type = null;
+        }
+
         if (!\is_array($values)) {
             $values = [$values];
         }
@@ -175,18 +197,26 @@ class SearchFilter extends AbstractContextAwareFilter implements SearchFilterInt
         $valueParameter = ':'.$queryNameGenerator->generateParameterName($field);
         $aliasedField = sprintf('%s.%s', $alias, $field);
 
-        if (null == $strategy || self::STRATEGY_EXACT == $strategy) {
-            if (1 == \count($values)) {
+        if (self::STRATEGY_EXACT === $strategy) {
+            if (1 === \count($values)) {
                 $queryBuilder
                     ->andWhere($queryBuilder->expr()->eq($wrapCase($aliasedField), $wrapCase($valueParameter)))
-                    ->setParameter($valueParameter, $values[0]);
+                    ->setParameter($valueParameter, $values[0], $type);
 
                 return;
             }
 
+            $parameters = $queryBuilder->getParameters();
+            $inQuery = [];
+            foreach ($values as $value) {
+                $inQuery[] = $valueParameter;
+                $parameters->add(new Parameter($valueParameter, $caseSensitive ? $value : strtolower($value), $type));
+                $valueParameter = ':'.$queryNameGenerator->generateParameterName($field);
+            }
+
             $queryBuilder
-                ->andWhere($queryBuilder->expr()->in($wrapCase($aliasedField), $valueParameter))
-                ->setParameter($valueParameter, $caseSensitive ? $values : array_map('strtolower', $values));
+                ->andWhere($wrapCase($aliasedField).' IN ('.implode(', ', $inQuery).')')
+                ->setParameters($parameters);
 
             return;
         }
@@ -228,7 +258,7 @@ class SearchFilter extends AbstractContextAwareFilter implements SearchFilterInt
         }
 
         $queryBuilder->andWhere($queryBuilder->expr()->orX(...$ors));
-        array_walk($parameters, [$queryBuilder, 'setParameter']);
+        array_walk($parameters, [$queryBuilder, 'setParameter'], $type);
     }
 
     /**