+++ src/Security/ResourceAccessChecker.php
@@ -15,6 +15,7 @@ namespace ApiPlatform\Core\Security;
 
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\Security\Core\Authentication\AuthenticationTrustResolverInterface;
+use Symfony\Component\Security\Core\Authentication\Token\NullToken;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
@@ -33,20 +34,31 @@ final class ResourceAccessChecker implements ResourceAccessCheckerInterface
     private $roleHierarchy;
     private $tokenStorage;
     private $authorizationChecker;
+    private $exceptionOnNoToken;
 
-    public function __construct(ExpressionLanguage $expressionLanguage = null, AuthenticationTrustResolverInterface $authenticationTrustResolver = null, RoleHierarchyInterface $roleHierarchy = null, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null)
+    public function __construct(ExpressionLanguage $expressionLanguage = null, AuthenticationTrustResolverInterface $authenticationTrustResolver = null, RoleHierarchyInterface $roleHierarchy = null, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null, bool $exceptionOnNoToken = true)
     {
         $this->expressionLanguage = $expressionLanguage;
         $this->authenticationTrustResolver = $authenticationTrustResolver;
         $this->roleHierarchy = $roleHierarchy;
         $this->tokenStorage = $tokenStorage;
         $this->authorizationChecker = $authorizationChecker;
+        $this->exceptionOnNoToken = $exceptionOnNoToken;
     }
 
     public function isGranted(string $resourceClass, string $expression, array $extraVariables = []): bool
     {
         if (null === $this->tokenStorage || null === $this->authenticationTrustResolver) {
             throw new \LogicException('The "symfony/security" library must be installed to use the "security" attribute.');
+        }
+        if (null === $token = $this->tokenStorage->getToken()) {
+            if ($this->exceptionOnNoToken) {
+                throw new \LogicException('The current token must be set to use the "security" attribute (is the URL behind a firewall?).');
+            }
+
+            if (class_exists(NullToken::class)) {
+                $token = new NullToken();
+            }
         }
         if (null === $this->expressionLanguage) {
             throw new \LogicException('The "symfony/expression-language" library must be installed to use the "security" attribute.');
@@ -57,7 +69,7 @@ final class ResourceAccessChecker implements ResourceAccessCheckerInterface
             'auth_checker' => $this->authorizationChecker, // needed for the is_granted expression function
         ]);
 
-        if ($token = $this->tokenStorage->getToken()) {
+        if ($token) {
             $variables = array_merge($variables, $this->getVariables($token));
         }